requirements.txt
llama_index
llama-index-llms-gemini
llama-index-embeddings-gemini
llama-index-vector-stores-mongodb
pymongo
langchain
langchain_core
langchain_google_genai
langchain_text_splitters
python-docx
pypdf2
